removed inner timeline; moved trips future into trip timeline; fixed edit_trip.dart;
