--- src/flutter_app/lib/components/diary_page/inner_timeline.dart
@@ -1,53 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:timelines/timelines.dart';
-
-class InnerTimeline extends StatelessWidget {
-  const InnerTimeline({super.key,
-    required this.spots,
-  });
-
-  final List<String> spots;
-
-  @override
-  Widget build(BuildContext context) {
-    bool isEdgeIndex(int index) {
-      return index == 0 || index == spots.length + 1;
-    }
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8.0),
-      child: FixedTimeline.tileBuilder(
-        theme: TimelineTheme.of(context).copyWith(
-          nodePosition: 0,
-          connectorTheme: TimelineTheme.of(context).connectorTheme.copyWith(
-            thickness: 1.0,
-          ),
-          indicatorTheme: TimelineTheme.of(context).indicatorTheme.copyWith(
-            size: 10.0,
-            position: 0.5,
-          ),
-        ),
-        builder: TimelineTileBuilder(
-          indicatorBuilder: (_, index) =>
-          !isEdgeIndex(index) ? Indicator.outlined(borderWidth: 1.0) : null,
-          startConnectorBuilder: (_, index) => Connector.solidLine(),
-          endConnectorBuilder: (_, index) => Connector.solidLine(),
-          contentsBuilder: (_, index) {
-            if (isEdgeIndex(index)) {
-              return null;
-            }
-
-            return Padding(
-              padding: const EdgeInsets.only(left: 8.0),
-              child: Text(spots[index - 1]),
-            );
-          },
-          itemExtentBuilder: (_, index) => isEdgeIndex(index) ? 10.0 : 30.0,
-          nodeItemOverlapBuilder: (_, index) =>
-          isEdgeIndex(index) ? true : null,
-          itemCount: spots.length + 2,
-        ),
-      ),
-    );
-  }
-}
